Fixed formatting so it works for any player num
first number is the number of players.
following 4 lines are restock data for each resource.
you can add data for all number of players by having a new line with
player num and the following 4 lines being data.
This is because the parser in the restockData method of ResourceMarket
automatically finds the one that matches the current number of players.